Change install.py to always set LD_RUN_PATH in cmd shell
The env var LD_RUN_PATH is set to ${prefix_arch}/lib and then exported as part
of the shell commands whenever bash build/install commands are executed.  This
is intended to reduce or eliminate the need for setting LD_LIBRARY_PATH.
Testing using NumPy showed this to be successful.