Remove gtk
This commit effectively reverts 0ea6144 06fa8da 021b8f0 and ab23d71 .
The the gtk build is removed from the default 64 bit make, and the custom
build of matplotlib to support the gtk libs is removed.  Specific
versions of freetype and libpng are no longer called out in pkgs.conda
to support a custom fontconfig build.