Set Crypt::SSLeay for noninteractive install
The --default option to Makefile.PL turns off asking about which SSL lib
to use (this seems OK for our purposes).  There was no Makefile.PL option
to turn off asking about live tests, and the asking part didn't seem to work
in the context of install.py (it was just hanging), so the pkgs.manifest
now references a custom package which just sets the test config to do
the tests without prompting.